✨ RENDERER: Refactor DOM Traversal
💡 **What**: Consolidated duplicated DOM traversal logic (`findAllImages`, `findAllScopes`, `findAllElements`) into `dom-scripts.ts` and updated `DomStrategy` (preload) and `SeekTimeDriver` to use shared constants.
🎯 **Why**: To reduce code duplication and ensure consistent Shadow DOM handling across the renderer package.
📊 **Impact**: Improves maintainability and reduces the risk of divergence between DOM discovery logic in different components.
🔬 **Verification**:
- `npx tsx packages/renderer/tests/verify-enhanced-dom-preload.ts` (Passed)
- `npm run render:canvas-example` (Passed)
