+++ docs/status/RENDERER.md
@@ -1,10 +1,11 @@
-**Version**: 1.77.0
+**Version**: 1.77.1
 
 **Posture**: MAINTENANCE WITH V2 EXPANSION
 
 # Renderer Agent Status
 
 ## Progress Log
+- [1.77.1] ✅ Completed: Refactor DOM Traversal - Consolidated duplicated DOM traversal logic (`findAllImages`, `findAllScopes`, `findAllElements`) into `dom-scripts.ts` and updated `DomStrategy` (preload) and `SeekTimeDriver` to use shared constants. Verified with `verify-enhanced-dom-preload.ts`.
 - [1.77.0] ✅ Completed: Abstraction for Pluggable Execution - Decoupled `RenderOrchestrator` from concrete `Renderer` implementation by introducing `RenderExecutor` interface and `LocalExecutor` strategy. Verified with `verify-orchestrator-executor.ts`.
 - [1.76.1] ✅ Completed: Restore Development Environment - Restored dependencies (`npm install`, `npx playwright install`) and fixed regressions in verification scripts (`verify-dom-preload.ts`, `verify-transparency.ts`), ensuring full test suite integrity.
 - [1.76.0] ✅ Completed: Canvas Image Preload - Implemented comprehensive asset preloading for `CanvasStrategy` (images, fonts, backgrounds) using a shared utility extracted from `DomStrategy`. Verified with `verify-canvas-preload.ts`.
